+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.1
+
+Fixed support for keyword parameters in `ObjectTools#apply` in Ruby 3.2.
+
 ## 1.1.0
 
 Files in the `Toolbox` are now autoloaded.